pyproject.toml
[project]
name = "finsight"
version = "0.1.0"
description = ""
authors = [
    {name = "lbatresm"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "langchain[openai] (>=0.3.27,<0.4.0)",
    "langchain-cohere (>=0.3,<0.4)",
    "langchain-openai (>=0.2,<0.3)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "langgraph (>=0.2.20,<0.3.0)",
    "playwright (>=1.55.0,<2.0.0)",
    "rich (>=14.1.0,<15.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
